requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
drf-yasg==1.21.10
ecdsa==0.19.1
inflection==0.5.1
MarkupSafe==3.0.2
packaging==25.0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
Werkzeug==3.1.3
